requirements.txt
openai
requests
Typer
pydantic
langchain
python-dotenv
